Allowed the case when the user is not signed in.

+++ app/assets/javascripts/projects.js.coffee
@@ -1,3 +1,4 @@
+# The view of the current active project.
 class ProjectView
   constructor: ->
     @currentTrialView = new window.CurrentTrialView
@@ -6,22 +7,23 @@ class ProjectView
       $(ui.panel).append @resultHtml
       @projectTabs.tabs 'select', "\##{ui.panel.id}"
 
+    onCloseTab = (target) =>
+      index = $('li', @projectTabs).index(target)
+      @projectTabs.tabs 'remove', index
+      trialId = target.find('a').attr('data-trial-href').split('-')[1]
+      if parseInt(trialId)
+        deleteUrl = "/trials/#{trialId}"
+        $.ajax({ dataType: 'json', type: 'DELETE', url: deleteUrl });
+
     @projectTabs = $('#current-project').tabs({
       tabTemplate: '''
                    <li>
-                     <a href="#{href}" data-trial-id="#{href}">#{label}</a> 
+                     <a href="#{href}" data-trial-href="#{href}">#{label}</a> 
                      <span class="ui-icon ui-icon-close">Remove Tab</span>
                    </li> 
                    ''',
       add: onAddTab
     })
-    
-    onCloseTab = (target) =>
-      index = $('li', @projectTabs).index(target)
-      @projectTabs.tabs 'remove', index
-      trial_id = target.find('a').attr('data-trial-id').split('-')[1]
-      deleteUrl = "/trials/#{trial_id}"
-      $.ajax({ dataType: 'json', type: 'DELETE', url: deleteUrl });
     
     $('#project-nav span.ui-icon-close').live 'click', 
                                               -> onCloseTab($(this).parent()) 
@@ -58,7 +60,11 @@ class ProjectView
 
     trialElement = $($(@resultHtml).filter('[data-trial-id]')[0])
     trialId = trialElement.attr('data-trial-id')
-    trialName = $.trim trialElement.text()                
+    trialName = $.trim trialElement.text()     
+    # If the user is not logged in, the trial id is -1 and the trial name should
+    # be unique, so we use the name as the id. 
+    if trialId is '-1'
+      trialId = trialName          
     @projectTabs.tabs 'add', "\#trial-#{trialId}", "#{trialName}", 1
 
 # ProjectController handles user interactions on the project view.
